Initial functions added
