+++ parameters.py
@@ -1,0 +1,7 @@
+pdfNumber = 1
+frameNumber = 0
+gridFrames = {}
+frameHighlighted = None
+pdfNames = {}
+row = 2
+column = 0
